[Refactor]: Passport에 @Parameter(hidden=true) 어노테이션 추가, Ticket 작성 로직 일부 수정, User 권한 체크 추가

+++ ticket/src/main/java/com/codeterian/ticket/presentation/controller/TicketController.java
@@ -72,7 +72,7 @@ public class TicketController {
     @GetMapping
     @Operation(summary = "티켓 전체 조회", description = "티켓 전체 조회 API")
     public ResponseEntity<ResponseDto<List<TicketFindResponseDto>>> ticketList(
-            @CurrentPassport Passport passport
+            @Parameter(hidden = true) @CurrentPassport Passport passport
     ) throws IllegalAccessException {
 
         if (passport.getUserRole()== UserRole.CUSTOMER) {
@@ -99,8 +99,15 @@ public class TicketController {
     @DeleteMapping("/{ticketId}")
     @Operation(summary = "티켓 삭제", description = "티켓 삭제 API")
     public ResponseEntity<ResponseDto<Void>> ticketDelete(
-            @PathVariable("ticketId") UUID ticketId
-    ) {
+            @PathVariable("ticketId") UUID ticketId,
+            @Parameter(hidden = true) @CurrentPassport Passport passport
+    ) throws IllegalAccessException {
+
+        if(passport.getUserRole()==UserRole.CUSTOMER) {
+            throw new IllegalAccessException();
+        }
+
+
         ticketService.deleteTicketById(ticketId);
 
         return ResponseEntity.ok(ResponseDto.ok());
